+++ Changelog.md
@@ -3,6 +3,7 @@
 ## 1.1
 
 ## 1.0.2
+- Do not show 'Kotlin not configured' during project gradle sync
 
 ### JS
 - Safe calls (`x?.let { it }`) are now inlined